# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Register API
  description: API for user management
  contact:
    name: Maik
    email: maikgafurtado@gmail.com
  version: 1.0.0
servers:
- url: https://localhost:8443/cadastroapi/actrl/api
  description: Local HTTPS server
...
paths:
...
  /actrl/api/users/update/email:
    put:
      tags:
      - User Resource
      summary: Updates user's email
      description: Initiates email update process by sending a confirmation code
      operationId: emailUpdate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailUpdateDto'
      responses:
        default:
          description: default response
          content:
            application/json: {}
  /actrl/api/users/register:
    post:
      tags:
      - User Resource
      summary: Register a new user
      description: Creates a new user with provided email, username, and password.
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /actrl/api/users/update/username:
    put:
      tags:
      - User Resource
      summary: Updates user's username
      description: Changes the username after authentication
      operationId: updateUsername
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsernameUpdateDto'
      responses:
        default:
          description: default response
          content:
            application/json: {}
  /actrl/api/users/validate/email:
    post:
      tags:
      - User Resource
      summary: Validates email update code
      description: Checks if the code for confirming email update is valid
      operationId: validateEmailUpdateCode
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateCodeDto'
      responses:
        default:
          description: default response
          content:
            application/json: {}
  /actrl/api/users/validate/register:
    post:
      tags:
      - User Resource
      summary: Validates registration code
      description: Checks the user code and removes the temporary token on success
      operationId: validateRegistrationCode
      parameters:
      - name: Authorization
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateCodeDto'
      responses:
        default:
          description: default response
          content:
            application/json: {}
components:
  schemas:
    ErrorResponse:
      required:
      - message
      - timestamp
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: Invalid credentials
        timestamp:
          type: string
          description: Timestamp of the error
          format: date-time
      description: Error response details
...
    UserResponseDto:
      required:
      - email
      - status
      type: object
      properties:
        email:
          type: string
          description: Email of the user.
          example: user@example.com
        status:
          type: string
          description: User status (e.g., ACTIVE, VALIDATING, etc.)
          example: ACTIVE
          enum:
          - VALIDATED
          - NOT_VALIDATED
          - AWAITING_EMAIL_VALIDATION
      description: DTO representing user details.
...
    EmailUpdateDto:
      required:
      - newEmail
      - password
      type: object
      properties:
        newEmail:
          type: string
          description: New email address
          example: newemail@example.com
        password:
          type: string
          description: Current password
          example: myPassword123
      description: DTO for updating email
    RegisterDto:
      required:
      - email
      - password
      - username
      type: object
      properties:
        username:
          type: string
          description: Username of the new user
          example: maik
        email:
          type: string
          description: Email of the new user
          example: email@example.com
        password:
          type: string
          description: Password for the new user
          example: '12345678'
      description: DTO used to register new users
    UsernameUpdateDto:
      required:
      - newUsername
      - password
      type: object
      properties:
        newUsername:
          type: string
          description: New username
          example: new_username
        password:
          type: string
          description: Current password
          example: user_password
      description: DTO for updating a user's username
    ValidateCodeDto:
      required:
      - code
      type: object
      properties:
        code:
          type: string
          description: Code sent to the user
          example: ABC123
      description: DTO to validate a code sent to the user